Factor out composed function + weight into own type
